Safari-611.1.10/LayoutTests/editing/async-clipboard/clipboard-write-in-copy-event-handler-in-subframe-expected.txt

This test verifies that clipboard items can be written while dispatching the 'copy' event. To manually run the test, copy the 'Copy me' text below.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyWrote is true
PASS successfullyParsed is true

TEST COMPLETE
